Removed Old Files
To maintain compatibility, the old files have been removed and the improved files updated to use the original name
